+++ src/server/.babelrc.js
@@ -1,8 +1,13 @@
 'use strict'
 
+const plugins = ['@babel/transform-runtime', 'transform-require-extensions']
+
+if (process.env.BABEL_ESM)
+  plugins.push(require('../../babelPluginAddBabelRuntimeFileExtensions.js'))
+
 module.exports = {
   comments: false,
-  plugins: ['@babel/transform-runtime', 'transform-require-extensions'],
+  plugins,
   presets: [
     [
       '@babel/env',