+++ src/universal/.babelrc.js
@@ -1,15 +1,18 @@
 'use strict'
 
+const plugins = [
+  ['@babel/proposal-class-properties', { loose: true }],
+  '@babel/transform-runtime',
+  'transform-require-extensions'
+]
+
+if (process.env.BABEL_ESM)
+  plugins.push(require('../../babelPluginAddBabelRuntimeFileExtensions.js'))
+
 module.exports = {
   comments: false,
-  plugins: [
-    ['@babel/proposal-object-rest-spread', { loose: true, useBuiltIns: true }],
-    ['@babel/proposal-class-properties', { loose: true }],
-    '@babel/transform-runtime',
-    'transform-require-extensions'
-  ],
+  plugins,
   presets: [
-    { plugins: ['transform-replace-object-assign'] },
     [
       '@babel/env',
       {
@@ -19,6 +22,6 @@ module.exports = {
         loose: true
       }
     ],
-    ['@babel/react', { useBuiltIns: true }]
+    '@babel/react'
   ]
 }